+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * github-233: [BUG] Missing Hadoop dependencies for S3, Delta, etc
 * github-235: Implement new `rest` hook with fine control
 * github-229: A build target should not fail if Impala "COMPUTE STATS" fails
+* github-236: 'copy' target should not apply output schema
 
 
 # Version 0.26.1 - 2022-08-03